second commit

+++ tekoapp/repositories/usertoken/delete.py
@@ -3,5 +3,9 @@ from tekoapp.repositories import usertoken
 
 def by_tokenstring(tokenstring):
     user_token = usertoken.find.by_tokenstring(tokenstring=tokenstring)
-    models.db.session.delete(user_token)
-    models.db.session.commit()
+    if user_token:
+        models.db.session.delete(user_token)
+        models.db.session.commit()
+        return True
+    else:
+        return False
